backend/requirements.txt
Django==5.0.2
opentelemetry-distro==0.47b0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation-logging==0.47b0
requests>=2.28.0
python-dateutil>=2.8.0
redis>=4.5.0